\name{ammiData}
\alias{ammiData}

\title{
  Dataset for example
}

\description{
    Simulated dataset in completely randomized design to illustrate the resources of the \code{ammiBayes} package.
}

\usage{
  data(ammiData)
}

\details{
  amb = Environment (4 environments)
  
  rep = Repetition (9 repetitions)
  
  gen = Genotype (6 genotypes)
  
  prod = Variabel response
}

\keyword{datasets}
