\name{summary.ammiBayes}
\alias{summary.ammiBayes}

\title{
  Summary Method for ammiBayes object
}

\description{
  Returns (and prints) a summary list for \code{ammiBayes} object.
}

\usage{
  \method{summary}{ammiBayes}(object, ...)
}

\arguments{
  \item{object}{A given object of the class \code{ammiBayes}.}
  \item{\dots}{Potential further arguments (required by generic).}
}

\author{
    Cristian Tiago Erazo Mendes (\email{tiagoerazo@gmail.com})\cr
    Carlos Pereira da Silva (\email{ccpsilva81@hotmail.com})\cr
    Julio Silvio de Sousa Bueno Filho (\email{jssbueno@ufla.br})\cr
    Fabio Mathias Correa (\email{f.correa@ru.ac.za})
}


\seealso{
  \code{\link[ammiBayes]{ammiBayes}}
}


\keyword{package}

