\name{summary.ammiBayes}
\alias{summary.ammiBayes}

\title{
  Summary Method for ammiBayes object
}

\description{
  Returns (and prints) a summary list for \code{ammiBayes} object.
}

\usage{
  \method{summary}{ammiBayes}(object, ...)
}

\arguments{
  \item{object}{A given object of the class \code{ammiBayes}.}
  \item{\dots}{Potential further arguments (required by generic).}
}

\references{
OLIVEIRA,L.A.; SILVA, C. P.; NUVUNGA, J. J.; SILVA, A. Q.; BALESTRE, M. Credible Intervals for Scores in the AMMI with Random Effects for Genotype. Crop Science, v. 55, p. 465-476, 2015.
doi:  https://doi.org/10.2135/cropsci2014.05.0369

SILVA, C. P.; OLIVEIRA, L. A.; NUVUNGA, J. J.; PAMPLONA, A. K. A.; BALESTRE, M. A Bayesian Shrinkage Approach for AMMI Models. 
Plos One, v. 10, p. e0131414, 2015. doi: https://doi.org/10.1371/journal.pone.0131414.
} 


\author{
        Luciano A. Oliveira \cr
        Carlos P. Silva \cr
        Cristian T. E. Mendes \cr
        Alessandra Q. Silva \cr
        Joel J. Nuvunga \cr
        Marcio Balestre \cr
        Julio S. S. Bueno-Filho \cr
		Fabio M. Correa
}


\seealso{
  \code{\link[ammiBayes]{ammiBayes}}
}


\keyword{package}

